# open_api_spec.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Pokemon Ranker
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:5000/pokeranker/api/v1.0
paths:
  /pokemon:
    get:
      summary: return list of all pokemon (includes rank info)
      operationId: get_pokemon_list
      tags:
        - pokemon
      responses:
        '200':
          description: pokemon list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PokemonList'
  /votes:
    get:
      summary: get list of all votes
      operationId: get_votes
      tags:
        - vote
      responses:
        '200':
          description: list of votes
          content:
            application.json:
              schema:
                $ref: '#/components/schemas/VoteList'
    post:
      summary: update ranking based on who the user voted for
      requestBody:
        description: Rank vote from user
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserVote'
      operationId: submit_vote
      tags:
        - vote
      responses:
        '201':
          description: ranking result created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vote'
  /votes/{vote_id}:
    get:
      summary: return status of a vote
      parameters:
        - in: path
          name: vote_id
          schema:
            type: integer
            format: int64
          required: true
          description: ID of the vote to get the status of
      operationId: get_vote_status
      tags:
        - vote
      responses:
        '200':
          description: Status of vote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vote'
  /pokemon/{pokemon_id}:
    get:
      summary: return information of specific pokemon
      parameters:
        - in: path
          name: pokemon_id
          schema:
            type: integer
            format: int64
          required: true
          description: ID of the pokemon to get the info of
      operationId: get_pokemon
      tags:
        - pokemon
      responses:
        '200':
          description: Info of specific pokemon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PokemonData'
  /pokemon/gen/{gen}:
    get:
      summary: return list of pokemon for specified generation
      parameters:
        - in: path
          name: gen
          schema:
            type: integer
            format: int64
          required: true
          description: Which generation to get pokemon from
      operationId:  get_pokemon_list_gen
      tags:
        - pokemon
      responses:
        '200':
          description: pokemon ranks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PokemonList'
  /ranks/match:
    get:
      summary: return a head to head matchup between any two pokemon
      operationId: get_match
      tags:
        - getmatch
      responses:
        '200':
          description: Matchup between two pokemon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchupData'
  /ranks/match/gen/{gen}:
    get:
      summary: return a head to head matchup of a specific generation
      parameters:
        - in: path
          name: gen
          schema:
            type: integer
          required: true
          description: Which generation to get ranks from
      operationId: get_match_gen
      tags:
        - getmatch
      responses:
        '200':
          description: Matchup between two pokemon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchupData'

components:
  schemas:

    PokemonList:
      type: array
      items:
        $ref: "#/components/schemas/PokemonData"
    VoteList:
      type: array
      items:
        $ref: "#/components/schemas/UserVote"
    UserVote:
      required:
        - winner_id
        - loser_id
      properties:
        winner_id:
          type: integer
          format: int64
        loser_id:
          type: integer
          format: int64
    Vote:
      required:
        - vote_id
        - status
        - winner_id
        - loser_id
      properties:
        vote_id:
          type: integer
          format: int64
        status:
          type: string
        winner_id:
          type: integer
          format: int64
        loser_id:
          type: integer
          format: int64
    PokemonData:
      required:
        - pokemon_id
        - name
        - gen
        - rank
        - genRank
        - elo
      properties:
        pokemon_id:
          type: integer
          format: int64
        name:
          type: string
        gen:
          type: integer
          format: int64
        rank:
          type: integer
          format: int64
        genRank:
          type: integer
          format: int64
        elo:
          type: integer
          format: int64
    MatchupData:
      required:
        - one
        - two
        - one_id
        - two_id
      properties:
        one:
          type: string
        two:
          type: string
        oneid:
          type: integer
        twoid:
          type: integer